requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
fastapi>=0.111.0
uvicorn>=0.30.0
pyyaml>=6.0
requests>=2.32.0
python-dotenv>=1.0.0
